Extract trace store purge helpers
